persist delta streaming on page refresh
