+++ server/controllers/account.js
@@ -1,20 +1,21 @@
 const models = require('../../db/models');
 const axios = require('axios');
 
+const deltaWatcher = require('../utils/deltaWatcher');
+
 
 module.exports.getOne = (req, res) => {
   console.log('Inside account.js Controller getOne() ');
+   
+  //start cursor 
+  deltaWatcher(req);
 
-  // when using Nylus call
-  const authString = 'Bearer ' + req.session.nylasToken;
-  axios.get(`https://api.nylas.com/account`, {
-    headers: { Authorization: authString }
-  }).then(response => {
-    res.send(response.data);
-  })
-  .catch(err => {
-    console.log("Error retrieving account info from Email API: ", err);
-  });
+  //send back account info
+  res.send(JSON.stringify({
+    id: req.session.accountId,
+    email_address: req.session.accountEmail,
+    nylasToken: req.session.nylasToken
+  }));
 
 };
